Move materials handlers to separate controller
