+++ spec/controllers/materials_controller_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe EventsController, type: :controller do
+RSpec.describe MaterialsController, type: :controller do
   let(:valid_attributes) { FactoryGirl.attributes_for(:event) }
   let(:file) do
     filepath = Rails.root.join('spec/testfiles/actual.pdf')
@@ -28,7 +28,7 @@ RSpec.describe EventsController, type: :controller do
     it "uploads a file to the event's root material directory" do
       mock_writing_to_filesystem do
         upload_file
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(File.exists?(File.join(@event.material_path, file.original_filename)))
         expect(flash[:notice]).to match(I18n.t(:success_message, scope: 'events.material_area'))
       end
@@ -39,7 +39,7 @@ RSpec.describe EventsController, type: :controller do
         subdir = "subdir"
         mkdir(subdir)
         upload_file(path: subdir)
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(File.exists?(File.join(@event.material_path, subdir, file.original_filename)))
         expect(flash[:notice]).to match(I18n.t(:success_message, scope: 'events.material_area'))
       end
@@ -50,7 +50,7 @@ RSpec.describe EventsController, type: :controller do
         upload_file
         not_a_directory = file.original_filename
         upload_file(path: not_a_directory)
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(File.exists?(File.join(@event.material_path, not_a_directory, file.original_filename))).to be false
         expect(flash[:alert]).to match(I18n.t(:download_file_not_found, scope: 'events.material_area'))
       end
@@ -60,7 +60,7 @@ RSpec.describe EventsController, type: :controller do
       mock_writing_to_filesystem do
         FileUtils.rm_rf(@event.material_path) if File.exists?(@event.material_path)
         upload_file
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(Dir.exists?(@event.material_path))
         expect(File.exists?(File.join(@event.material_path, file.original_filename)))
         expect(flash[:notice]).to match(I18n.t(:success_message, scope: 'events.material_area'))
@@ -70,7 +70,7 @@ RSpec.describe EventsController, type: :controller do
     it "shows error if no file was given" do
       mock_writing_to_filesystem do
         post :upload_material, event_id: @event.to_param
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(flash[:alert]).to match(I18n.t(:no_file_given, scope: 'events.material_area'))
       end
     end
@@ -79,7 +79,7 @@ RSpec.describe EventsController, type: :controller do
       mock_writing_to_filesystem do
         allow(File).to receive(:write).and_raise(IOError)
         upload_file
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(flash[:alert]).to match(I18n.t(:saving_fails, scope: 'events.material_area'))
       end
     end
@@ -89,7 +89,7 @@ RSpec.describe EventsController, type: :controller do
         dangerous_name = "../hacked_you.lol"
         allow_any_instance_of(Rack::Test::UploadedFile).to receive(:original_filename).and_return(dangerous_name)
         upload_file
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(File.exists?(File.join(@event.material_path, file.original_filename))).to be false
         expect(flash[:alert]).to match(I18n.t(:invalid_path_given, scope: 'events.material_area'))
       end
@@ -101,7 +101,7 @@ RSpec.describe EventsController, type: :controller do
       mock_writing_to_filesystem do
         dirname = "testdir"
         mkdir(dirname)
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(Dir.exists?(File.join(@event.material_path, dirname)))
         expect(flash[:notice]).to match(I18n.t(:dir_created, scope: 'events.material_area'))
       end
@@ -112,7 +112,7 @@ RSpec.describe EventsController, type: :controller do
         FileUtils.rm_rf(@event.material_path) if File.exists?(@event.material_path)
         subdir = "subdir"
         mkdir(subdir)
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(Dir.exists?(@event.material_path))
         expect(Dir.exists?(File.join(@event.material_path, subdir)))
         expect(flash[:notice]).to match(I18n.t(:dir_created, scope: 'events.material_area'))
@@ -123,7 +123,7 @@ RSpec.describe EventsController, type: :controller do
       mock_writing_to_filesystem do
         dangerous_name = "../hacked_you.lol"
         mkdir(dangerous_name)
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(File.exists?(File.join(@event.material_path, dangerous_name))).to be false
         expect(flash[:alert]).to match(I18n.t(:invalid_path_given, scope: 'events.material_area'))
       end
@@ -132,7 +132,7 @@ RSpec.describe EventsController, type: :controller do
     it "does not accept unset parameters" do
       mock_writing_to_filesystem do
         post :make_material_folder, event_id: @event.to_param, path: ''
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(flash[:alert]).to match(I18n.t(:no_file_given, scope: 'events.material_area'))
       end
     end
@@ -142,7 +142,7 @@ RSpec.describe EventsController, type: :controller do
         name = "subdir"
         dangerous_path = "../hacked_you.lol"
         mkdir(name, path: dangerous_path)
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(flash[:alert]).to match(I18n.t(:invalid_path_given, scope: 'events.material_area'))
       end
     end
@@ -154,7 +154,7 @@ RSpec.describe EventsController, type: :controller do
         mkdir(name)
         upload_file(name: filename, path: name) # empty file to test if first dir was deleted
         mkdir(name)
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(File.exists?(File.join(@event.material_path, name, "gentoo.txt")))
         expect(flash[:alert]).to match(I18n.t(:already_exists, scope: 'events.material_area'))
       end
@@ -165,7 +165,7 @@ RSpec.describe EventsController, type: :controller do
         name = "gentoo.txt"
         upload_file(name: name)
         mkdir(name)
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(File.directory?(File.join(@event.material_path, name))).to be false
         expect(flash[:alert]).to match(I18n.t(:already_exists, scope: 'events.material_area'))
       end
@@ -176,7 +176,7 @@ RSpec.describe EventsController, type: :controller do
         name = "mydir"
         path = "myunexistingdir"
         mkdir(name, path: path)
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(File.directory?(File.join(@event.material_path, path, name))).to be false
         expect(flash[:alert]).to match(I18n.t(:download_file_not_found, scope: 'events.material_area'))
       end
@@ -189,7 +189,7 @@ RSpec.describe EventsController, type: :controller do
         path = "gentoo.txt"
         upload_file(name: path)
         remove_file(path)
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(File.exists?(File.join(@event.material_path, path))).to be false
         expect(flash[:notice]).to match(I18n.t(:file_removed, scope: 'events.material_area'))
       end
@@ -198,7 +198,7 @@ RSpec.describe EventsController, type: :controller do
     it "does not accept unset parameters" do
       mock_writing_to_filesystem do
         post :remove_material, event_id: @event.to_param
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(flash[:alert]).to match(I18n.t(:no_file_given, scope: 'events.material_area'))
       end
     end
@@ -208,7 +208,7 @@ RSpec.describe EventsController, type: :controller do
         path = "darjeeling"
         mkdir(path)
         remove_file(path)
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(File.exists?(File.join(@event.material_path, path))).to be false
         expect(flash[:notice]).to match(I18n.t(:file_removed, scope: 'events.material_area'))
       end
@@ -218,7 +218,7 @@ RSpec.describe EventsController, type: :controller do
       mock_writing_to_filesystem do
         dangerous_path = "../your_root_directory"
         remove_file(dangerous_path)
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(flash[:alert]).to match(I18n.t(:invalid_path_given, scope: 'events.material_area'))
       end
     end
@@ -227,7 +227,7 @@ RSpec.describe EventsController, type: :controller do
       mock_writing_to_filesystem do
         path = "photo_of_my_girlfriend.jpg"
         remove_file(path)
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(flash[:alert]).to match(I18n.t(:download_file_not_found, scope: 'events.material_area'))
       end
     end
@@ -240,7 +240,7 @@ RSpec.describe EventsController, type: :controller do
         mkdir(subdir)
         upload_file
         move_file(file.original_filename, subdir)
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(File.exists?(File.join(@event.material_path, subdir, file.original_filename)))
         expect(flash[:notice]).to match(I18n.t(:file_moved, scope: 'events.material_area'))
       end
@@ -252,7 +252,7 @@ RSpec.describe EventsController, type: :controller do
         dangerous_path = "../your_root_directory"
         path = "myfile.txt"
         move_file(dangerous_path, path)
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(flash[:alert]).to match(I18n.t(:invalid_path_given, scope: 'events.material_area'))
       end
     end
@@ -264,7 +264,7 @@ RSpec.describe EventsController, type: :controller do
         mkdir(subdir)
         mkdir(name)
         move_file(name, subdir)
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(Dir.exists?(File.join(@event.material_path, subdir, name)))
         expect(flash[:notice]).to match(I18n.t(:file_moved, scope: 'events.material_area'))
       end
@@ -273,7 +273,7 @@ RSpec.describe EventsController, type: :controller do
     it "does not accept unset parameters" do
       mock_writing_to_filesystem do
         post :move_material, event_id: @event.to_param
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(flash[:alert]).to match(I18n.t(:no_file_given, scope: 'events.material_area'))
       end
     end
@@ -292,7 +292,7 @@ RSpec.describe EventsController, type: :controller do
         mkdir(subdir)
         upload_file(name: file.original_filename, path: subdir, content: text)
         move_file(file.original_filename, subdir)
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(File.exists?(File.join(@event.material_path, subdir, file.original_filename)))
         expect(File.read(File.join(@event.material_path, subdir, file.original_filename))).to eq text
         expect(flash[:alert]).to match(I18n.t(:already_exists, scope: 'events.material_area'))
@@ -304,7 +304,7 @@ RSpec.describe EventsController, type: :controller do
         subdir = "subdir"
         mkdir(subdir)
         move_file(file.original_filename, subdir)
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(File.exists?(File.join(@event.material_path, subdir, file.original_filename))).to be false
         expect(flash[:alert]).to match(I18n.t(:download_file_not_found, scope: 'events.material_area'))
       end
@@ -316,7 +316,7 @@ RSpec.describe EventsController, type: :controller do
         subdir = "subdir"
         upload_file
         move_file(file.original_filename, subdir)
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(File.exists?(File.join(@event.material_path, subdir, file.original_filename))).to be false
         expect(flash[:alert]).to match(I18n.t(:download_file_not_found, scope: 'events.material_area'))
       end
@@ -327,7 +327,7 @@ RSpec.describe EventsController, type: :controller do
         dir = "dir"
         mkdir(dir)
         move_file(dir, dir)
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(File.exists?(File.join(@event.material_path, dir)))
         expect(flash[:alert]).to match(I18n.t(:cant_move_in_child, scope: 'events.material_area'))
       end
@@ -340,7 +340,7 @@ RSpec.describe EventsController, type: :controller do
         mkdir(dir)
         mkdir(subdir, path: dir)
         move_file(dir, [dir, subdir])
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(File.exists?(File.join(@event.material_path, dir)))
         expect(File.exists?(File.join(@event.material_path, dir, subdir)))
         expect(flash[:alert]).to match(I18n.t(:cant_move_in_child, scope: 'events.material_area'))
@@ -354,7 +354,7 @@ RSpec.describe EventsController, type: :controller do
         new_name = "otter_in_my_water"
         upload_file
         rename(file.original_filename, new_name)
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(File.exists?(File.join(@event.material_path, file.original_filename))).to be false
         expect(File.exists?(File.join(@event.material_path, new_name)))
         expect(flash[:notice]).to match(I18n.t(:file_renamed, scope: 'events.material_area'))
@@ -367,7 +367,7 @@ RSpec.describe EventsController, type: :controller do
         new_name = "meguca"
         mkdir(name)
         rename(name, new_name)
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(Dir.exists?(File.join(@event.material_path, new_name)))
         expect(Dir.exists?(File.join(@event.material_path, name))).to be false
         expect(flash[:notice]).to match(I18n.t(:file_renamed, scope: 'events.material_area'))
@@ -377,7 +377,7 @@ RSpec.describe EventsController, type: :controller do
     it "does not accept unset parameters" do
       mock_writing_to_filesystem do
         post :rename_material, event_id: @event.to_param
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(flash[:alert]).to match(I18n.t(:no_file_given, scope: 'events.material_area'))
       end
     end
@@ -388,7 +388,7 @@ RSpec.describe EventsController, type: :controller do
         to = 'dir/test'
         mkdir('dir')
         rename(invalid_from, to)
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(File.exists?(File.join(@event.material_path, to))).to be false
         expect(flash[:alert]).to match(I18n.t(:invalid_path_given, scope: 'events.material_area'))
       end
@@ -400,7 +400,7 @@ RSpec.describe EventsController, type: :controller do
         invalid_to = '../../../test/invalid'
         mkdir('dir')
         rename(from, invalid_to)
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(File.exists?(File.join(@event.material_path, invalid_to))).to be false
         expect(flash[:alert]).to match(I18n.t(:invalid_path_given, scope: 'events.material_area'))
       end
@@ -412,7 +412,7 @@ RSpec.describe EventsController, type: :controller do
         invalid_to = 'test/invalid'
         mkdir('dir')
         rename(from, invalid_to)
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(File.exists?(File.join(@event.material_path, invalid_to))).to be false
         expect(flash[:alert]).to match(I18n.t(:invalid_path_given, scope: 'events.material_area'))
       end
@@ -422,7 +422,7 @@ RSpec.describe EventsController, type: :controller do
       mock_writing_to_filesystem do
         upload_file
         rename(file.original_filename, file.original_filename)
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(File.exists?(File.join(@event.material_path, file.original_filename)))
       end
     end
@@ -434,7 +434,7 @@ RSpec.describe EventsController, type: :controller do
         upload_file
         upload_file(name: first_file, content: text)
         rename(file.original_filename, first_file)
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(File.exists?(File.join(@event.material_path, file.original_filename)))
         expect(File.exists?(File.join(@event.material_path, first_file)))
         expect(File.read(File.join(@event.material_path, first_file))).to eq text
@@ -446,7 +446,7 @@ RSpec.describe EventsController, type: :controller do
       mock_writing_to_filesystem do
         new_name = "serenity"
         rename(file.original_filename, new_name)
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(File.exists?(File.join(@event.material_path, new_name))).to be false
         expect(flash[:alert]).to match(I18n.t(:download_file_not_found, scope: 'events.material_area'))
       end
@@ -493,7 +493,7 @@ RSpec.describe EventsController, type: :controller do
       mock_writing_to_filesystem do
         upload_file
         post :download_material, event_id: @event.to_param
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(flash[:alert]).to match(I18n.t(:no_file_given, scope: 'events.material_area'))
       end
     end
@@ -503,7 +503,7 @@ RSpec.describe EventsController, type: :controller do
         filename = "non-existing/file.pdf"
         upload_file
         download_file(filename)
-        expect(response).to redirect_to :action => :show, :id => @event.id
+        expect(response).to redirect_to :controller => :events, :action => :show, :id => @event.id
         expect(flash[:alert]).to match(I18n.t(:download_file_not_found, scope: 'events.material_area'))
       end
     end
